PaymentInstruction CRUD

+++ src/main/java/com/mdbprocessor/persistence/PaymentInstruction.java
@@ -1,16 +1,14 @@
 package com.mdbprocessor.persistence;
 
 import lombok.Data;
-
 import javax.persistence.*;
 
-//persist
 @Data
 @Entity
 public class PaymentInstruction {
     @Id
     @GeneratedValue
-    private long id;
+    private int id;
 
     @Column(name = "message_id")
     private String message_id;
